split SimpleCraftCommands into Client and Server parts
